+++ src/components/FLToWA/Florida.tsx
@@ -1,0 +1,22 @@
+export function Florida() {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      version="1.0"
+      width="100.000000pt"
+      height="100.000000pt"
+      viewBox="0 0 100.000000 100.000000"
+      preserveAspectRatio="xMidYMid meet"
+    >
+      <g
+        transform="translate(0.000000,100.000000) scale(0.100000,-0.100000)"
+        fill="#FFFFFF"
+        stroke="none"
+      >
+        <path d="M4 945 c-4 -9 -1 -28 4 -43 6 -15 12 -38 14 -52 3 -27 12 -28 104 -14 34 5 49 1 85 -22 24 -15 53 -38 64 -51 31 -35 119 -25 158 19 15 18 20 18 35 7 24 -19 52 -47 52 -54 0 -10 38 -43 57 -49 32 -10 47 -49 34 -86 -17 -51 -3 -134 30 -177 16 -20 34 -54 41 -75 6 -22 20 -43 30 -47 10 -4 26 -22 36 -41 10 -22 26 -36 39 -38 16 -2 25 -12 29 -30 7 -36 17 -54 37 -66 15 -10 21 -8 97 28 30 14 41 54 47 169 l4 87 -45 86 c-30 57 -44 93 -40 108 3 15 -6 36 -31 70 -35 45 -85 150 -85 178 0 7 -5 29 -10 49 -9 30 -16 38 -51 48 -37 11 -43 11 -63 -8 -20 -19 -34 -21 -164 -21 -131 0 -142 1 -152 20 -10 19 -21 20 -181 20 -142 0 -171 -2 -175 -15z m336 -45 c10 -19 21 -20 174 -20 158 0 164 -1 169 -21 8 -32 32 -17 32 20 0 54 28 33 46 -34 23 -81 54 -151 85 -187 15 -19 22 -37 19 -48 -4 -11 14 -55 45 -112 47 -90 50 -99 50 -171 0 -42 -4 -79 -10 -82 -5 -3 -10 -17 -10 -31 0 -15 -7 -27 -17 -30 -10 -3 -26 -10 -35 -15 -25 -13 -26 -13 -33 30 -5 32 -13 42 -41 56 -21 10 -34 23 -34 35 0 10 -11 27 -25 38 -14 11 -38 45 -54 75 l-29 56 24 16 c29 19 31 46 4 70 -15 14 -23 16 -32 7 -8 -8 -8 -15 0 -25 7 -8 9 -17 5 -20 -24 -25 -37 28 -22 88 10 40 9 51 -7 81 -11 19 -29 37 -41 41 -13 3 -23 8 -23 12 -1 11 -64 77 -92 94 -37 23 -79 22 -85 -2 -4 -17 -41 -41 -63 -41 -5 0 -43 23 -83 50 -60 40 -82 50 -116 50 -23 0 -51 -5 -62 -10 -15 -8 -19 -7 -19 4 0 8 -3 21 -6 30 -5 14 10 16 135 16 130 0 141 -1 151 -20z" />
+        <path d="M973 123 c-9 -27 -45 -64 -72 -74 -17 -7 -21 -13 -14 -20 23 -23 113 59 113 101 0 17 -20 11 -27 -7z" />
+        <path d="M803 26 c-34 -16 -28 -29 10 -21 41 9 49 14 42 26 -8 11 -23 10 -52 -5z" />
+      </g>
+    </svg>
+  );
+}
