+++ src/components/FLToWA/Washington.tsx
@@ -1,0 +1,21 @@
+export function Washington() {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      version="1.0"
+      width="100.000000pt"
+      height="100.000000pt"
+      viewBox="0 0 100.000000 100.000000"
+      preserveAspectRatio="xMidYMid meet"
+    >
+      <g
+        transform="translate(0.000000,100.000000) scale(0.100000,-0.100000)"
+        fill="#FFFFFF"
+        stroke="none"
+      >
+        <path d="M249 819 c-9 -18 -7 -26 10 -50 22 -29 29 -99 11 -99 -5 0 -10 9 -10 19 0 31 -19 35 -52 11 -36 -27 -78 -22 -133 16 -21 14 -45 24 -54 22 -22 -4 -27 -89 -7 -120 8 -13 18 -43 22 -68 4 -25 11 -54 17 -65 6 -11 10 -52 9 -92 0 -56 4 -78 18 -95 15 -19 20 -21 34 -9 13 11 22 11 39 3 12 -5 29 -13 39 -15 10 -4 19 -22 23 -47 6 -36 12 -44 42 -57 39 -16 78 -11 105 14 12 11 25 13 45 8 51 -14 147 -14 170 1 12 8 28 11 35 9 8 -3 34 2 58 12 31 11 88 18 185 22 l140 6 -4 25 c-2 14 -6 147 -10 295 l-6 270 -357 3 c-356 2 -357 2 -369 -19z m691 -248 c0 -126 3 -243 6 -260 l7 -31 -110 0 c-82 0 -124 -5 -165 -19 -38 -13 -59 -16 -66 -9 -7 7 -20 3 -41 -13 -25 -19 -33 -20 -46 -9 -12 10 -20 11 -38 1 -17 -9 -30 -9 -51 -2 -45 18 -86 13 -105 -11 -26 -31 -67 -19 -76 24 -4 18 -14 41 -23 51 -24 27 -79 50 -107 45 -25 -5 -34 5 -15 17 14 9 13 48 -2 63 -9 9 -8 12 4 12 30 0 32 23 4 45 -20 16 -31 37 -37 70 -5 26 -16 61 -24 78 -8 17 -15 38 -15 47 0 15 6 14 47 -7 37 -19 59 -24 96 -20 41 4 53 0 78 -21 26 -21 33 -23 56 -13 20 9 24 16 18 34 -4 12 -10 42 -12 67 -3 24 -12 55 -20 67 l-15 23 326 0 326 0 0 -229z" />
+        <path d="M143 788 c-28 -14 -29 -22 -3 -48 27 -27 39 -25 51 7 18 48 -1 64 -48 41z" />
+      </g>
+    </svg>
+  );
+}
