+++ src/components/index.ts
@@ -1,7 +1,9 @@
 export * from "./base";
 export * from "./Avatar";
 export * from "./CareerExperience";
+export * from "./ContactForm";
 export * from "./ContentCard";
+export * from "./FLToWA";
 export * from "./Footer";
 export * from "./Layout";
 export * from "./NavigationMenu";
