+++ src/pages/index.ts
@@ -1,4 +1,5 @@
 export * from "./About";
 export * from "./Career";
 export * from "./Contact";
+export * from "./Error";
 export * from "./Home";
