+++ tsconfig.json
@@ -16,5 +16,5 @@
     "noEmit": true,
     "jsx": "react-jsx"
   },
-  "include": ["src/*"]
+  "include": ["src/*", "src/components/App"]
 }
